Fixed "Close" distance mode
Tested Steven's code on the FPGA, and had to make one change to the code to make it work. The "distance <=" assignment was left outside the if-else structure, so data was taken from v2d_LUT regardless of the mux_close bit. Changed the distance assignment to be in the if-else structure.

Also added "Lab 3 Handout" file to the submissions folder for convenience.