Pre-lab submitted
I have submitted the pre-lab to D2L, and added the submissions to the "Submissions" folder. I have also added the file "binary_bcd_2stage.vhd" which is a copy of the modified version. Since it does not work on the FPGA I'll leave it there just as a backup.