Checkpoint for bonus
I have partially completed the bonus. I have it working to the point where the user can arbitrarily define the number of stages they want in the synchronizer. I had a hard time implementing using a generic variable/generate function, which I will work on now. If I am not able to figure it out we will use this version.