Started Lab 4
I wanted to get an early start on Lab 4 since I had some time today. Compared to Lab 3, I have created "distance2downcount.vhd" and "pwm_dac.vhd". The distance2downcount is something I wrote which scales the clock frequency down depending on the distance measured by the sensor. This then sends that scaled clock to pwm_dac which will output a pwm signal of an audible frequency. This is the first time I've gotten the code to compile, so this is just a checkpoint; more to follow.