It works! (Lab 4)
This morning I joined the office hours, and Milad helped me walk through the code and fix the issue. To be honest, I'm not sure what we changed that eventually worked, but we got it working is the important part! The TA also showed me a great tool called the "signal tap analyzer" which allows you to see a signal changing live while the device is running. This is what we used to confirm our code was working, and from there it was just a matter of figuring out pin assignments. I believe the issue was that I was using a the wrong pin on the Arduino header. Another possible issue was that I had included Arduino_IO[ ] pins in voltmeter.vhd. Either way, I just ran it and it's working great!